Fix dataset usage. (Partial revert of 8302996)
This undoes the changes in 8302996 to Timeseries so that we go back to using
Datasets as a context manager. To work-around the bug with `self.z` that is now
being read from the local Dataset instance wherever it is needed.

Timeseries `ncd` is deprecated in favor of `dataset` which is returned as a
context manager. The choice was made to return an opened dataset from `ncd`
as well, since this is actually the current behavior from before `close()`
was fixed.

Finally, update the URN tests to use the context manager properly. This is
required to ensure the dataset is closed before the next `add_variable` call.